refactor: product service 로직 분리 및 클래스 형식 통일
