+++ src/main/java/umc/stockoneqback/product/controller/ProductApiController.java
@@ -8,13 +8,10 @@ import umc.stockoneqback.global.base.BaseResponse;
 import umc.stockoneqback.global.base.GlobalErrorCode;
 import umc.stockoneqback.product.dto.request.EditProductRequest;
 import umc.stockoneqback.product.dto.response.GetRequiredInfoResponse;
-import umc.stockoneqback.product.dto.response.GetTotalProductResponse;
 import umc.stockoneqback.product.dto.response.LoadProductResponse;
-import umc.stockoneqback.product.dto.response.SearchProductResponse;
 import umc.stockoneqback.product.service.ProductService;
 
 import java.io.IOException;
-import java.util.List;
 
 @RestController
 @RequiredArgsConstructor
@@ -43,14 +40,6 @@ public class ProductApiController {
         return new BaseResponse<>(productService.loadProduct(userId, productId));
     }
 
-    @GetMapping("/search")
-    public BaseResponse<List<SearchProductResponse>> searchProduct(@ExtractPayload Long userId,
-                                                                   @RequestParam(value = "store") Long storeId,
-                                                                   @RequestParam(value = "condition") String storeConditionValue,
-                                                                   @RequestParam(value = "name") String productName) throws IOException {
-        return new BaseResponse<>(productService.searchProduct(userId, storeId, storeConditionValue, productName));
-    }
-
     @PatchMapping("/edit/{productId}")
     public BaseResponse<GlobalErrorCode> editProduct(@ExtractPayload Long userId,
                                                      @PathVariable(value = "productId") Long productId,
@@ -65,22 +54,5 @@ public class ProductApiController {
                                                        @PathVariable(value = "productId") Long productId) {
         productService.deleteProduct(userId, productId);
         return new BaseResponse<>(GlobalErrorCode.SUCCESS);
-    }
-
-    @GetMapping("/count")
-    public BaseResponse<List<GetTotalProductResponse>> getTotalProduct(@ExtractPayload Long userId,
-                                                                       @RequestParam(value = "store") Long storeId,
-                                                                       @RequestParam(value = "condition") String storeConditionValue) {
-        return new BaseResponse<>(productService.getTotalProduct(userId, storeId, storeConditionValue));
-    }
-
-    @GetMapping("/page")
-    public BaseResponse<List<SearchProductResponse>> getListOfSearchConditionProduct(@ExtractPayload Long userId,
-                                                                                     @RequestParam(value = "store") Long storeId,
-                                                                                     @RequestParam(value = "condition") String storeConditionValue,
-                                                                                     @RequestParam(value = "search") String searchConditionValue,
-                                                                                     @RequestParam(value = "last", defaultValue = "-1", required = false) Long productId,
-                                                                                     @RequestParam(value = "sort") String sortConditionValue) throws IOException {
-        return new BaseResponse<>(productService.getListOfSearchProduct(userId, storeId, storeConditionValue, searchConditionValue, productId, sortConditionValue));
     }
 }
