+++ src/main/java/umc/stockoneqback/product/infra/query/ProductFindQueryRepository.java
@@ -3,15 +3,15 @@ package umc.stockoneqback.product.infra.query;
 import umc.stockoneqback.product.domain.SearchCondition;
 import umc.stockoneqback.product.domain.SortCondition;
 import umc.stockoneqback.product.domain.StoreCondition;
-import umc.stockoneqback.product.infra.query.dto.FindProductPage;
+import umc.stockoneqback.product.infra.query.dto.ProductFindPage;
 import umc.stockoneqback.role.domain.store.Store;
 
 import java.util.List;
 
-public interface FindProductQueryRepository {
-    List<FindProductPage> findProductByName(Store store, StoreCondition storeCondition, String productName);
+public interface ProductFindQueryRepository {
+    List<ProductFindPage> findProductByName(Store store, StoreCondition storeCondition, String productName);
 
-    List<FindProductPage> findPageOfSearchConditionOrderBySortCondition
+    List<ProductFindPage> findPageOfSearchConditionOrderBySortCondition
             (Store store, StoreCondition storeCondition, SearchCondition searchCondition,
              SortCondition sortCondition, String productName, Long orderFreq, Integer pageSize);
 }
