+++ src/main/java/umc/stockoneqback/product/infra/query/dto/ProductFindPage.java
@@ -4,14 +4,14 @@ import com.querydsl.core.annotations.QueryProjection;
 import lombok.Getter;
 
 @Getter
-public class FindProductPage {
+public class ProductFindPage {
     private final Long id;
     private final String name;
     private final String imageUrl;
     private final Long stockQuant;
 
     @QueryProjection
-    public FindProductPage(Long id, String name, String imageUrl, Long stockQuant) {
+    public ProductFindPage(Long id, String name, String imageUrl, Long stockQuant) {
         this.id = id;
         this.name = name;
         this.imageUrl = imageUrl;
