+++ src/main/java/umc/stockoneqback/product/service/ProductFindOthersService.java
@@ -6,7 +6,7 @@ import org.springframework.transaction.annotation.Transactional;
 import umc.stockoneqback.product.domain.*;
 import umc.stockoneqback.product.dto.response.GetTotalProductResponse;
 import umc.stockoneqback.product.dto.response.SearchProductOthersResponse;
-import umc.stockoneqback.product.infra.query.dto.FindProductPage;
+import umc.stockoneqback.product.infra.query.dto.ProductFindPage;
 import umc.stockoneqback.role.domain.store.Store;
 
 import java.io.IOException;
@@ -19,41 +19,42 @@ import java.util.List;
 public class ProductFindOthersService {
     private final ProductRepository productRepository;
     private final ProductService productService;
+    private final ProductFindService productFindService;
     private static final Integer PAGE_SIZE = 9;
 
     @Transactional
     public List<SearchProductOthersResponse> searchProductOthers
             (Store store, String storeConditionValue, String productName) throws IOException {
         StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
-        List<FindProductPage> searchProductUrlList = productService.findProductAllByName(store, storeCondition, productName);
-        return convertUrlToResponse(searchProductUrlList);
+        List<ProductFindPage> searchProductUrlList = productFindService.findProductAllByName(store, storeCondition, productName);
+        return convertUrlToResponseOthers(searchProductUrlList);
     }
 
     @Transactional
     public List<GetTotalProductResponse> getTotalProductOthers(Store store, String storeConditionValue) {
         StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
-        return productService.countProduct(store, storeCondition);
+        return productFindService.countProduct(store, storeCondition);
     }
 
     @Transactional
     public List<SearchProductOthersResponse> getListOfSearchProductOthers
             (Store store, String storeConditionValue, String searchConditionValue, Long productId) throws IOException {
-        Product product = productService.configPaging(productId);
+        Product product = productFindService.configPaging(productId);
         StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
         SearchCondition searchCondition = SearchCondition.findSearchConditionByValue(searchConditionValue);
-        List<FindProductPage> searchProductUrlList = productRepository.findPageOfSearchConditionOrderBySortCondition
+        List<ProductFindPage> searchProductUrlList = productRepository.findPageOfSearchConditionOrderBySortCondition
                 (store, storeCondition, searchCondition, SortCondition.NAME, product.getName(), product.getOrderFreq(), PAGE_SIZE);
-        return convertUrlToResponse(searchProductUrlList);
+        return convertUrlToResponseOthers(searchProductUrlList);
     }
 
-    private List<SearchProductOthersResponse> convertUrlToResponse(List<FindProductPage> searchProductUrlList) throws IOException {
+    private List<SearchProductOthersResponse> convertUrlToResponseOthers(List<ProductFindPage> searchProductUrlList) throws IOException {
         List<SearchProductOthersResponse> searchProductOthersResponseList = new ArrayList<>();
-        for (FindProductPage findProductPage : searchProductUrlList) {
-            byte[] image = productService.getImageOrElseNull(findProductPage.getImageUrl());
+        for (ProductFindPage productFindPage : searchProductUrlList) {
+            byte[] image = productService.getImageOrElseNull(productFindPage.getImageUrl());
             SearchProductOthersResponse searchProductOthersResponse = SearchProductOthersResponse.builder()
-                    .id(findProductPage.getId())
-                    .name(findProductPage.getName())
-                    .stockQuant(findProductPage.getStockQuant())
+                    .id(productFindPage.getId())
+                    .name(productFindPage.getName())
+                    .stockQuant(productFindPage.getStockQuant())
                     .image(image)
                     .build();
             searchProductOthersResponseList.add(searchProductOthersResponse);
