+++ src/main/java/umc/stockoneqback/product/service/ProductService.java
@@ -10,13 +10,12 @@ import umc.stockoneqback.auth.service.TokenService;
 import umc.stockoneqback.file.service.FileService;
 import umc.stockoneqback.global.base.BaseException;
 import umc.stockoneqback.global.base.GlobalErrorCode;
-import umc.stockoneqback.product.domain.*;
+import umc.stockoneqback.product.domain.Product;
+import umc.stockoneqback.product.domain.ProductRepository;
+import umc.stockoneqback.product.domain.StoreCondition;
 import umc.stockoneqback.product.dto.response.GetRequiredInfoResponse;
-import umc.stockoneqback.product.dto.response.GetTotalProductResponse;
 import umc.stockoneqback.product.dto.response.LoadProductResponse;
-import umc.stockoneqback.product.dto.response.SearchProductResponse;
 import umc.stockoneqback.product.exception.ProductErrorCode;
-import umc.stockoneqback.product.infra.query.dto.FindProductPage;
 import umc.stockoneqback.role.domain.store.Store;
 import umc.stockoneqback.role.service.PartTimerService;
 import umc.stockoneqback.role.service.StoreService;
@@ -27,7 +26,6 @@ import umc.stockoneqback.user.service.UserFindService;
 
 import java.io.IOException;
 import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
 
@@ -42,7 +40,6 @@ public class ProductService {
     private final UserFindService userFindService;
     private final PartTimerService partTimerService;
     private final PassProductFCMService passProductFCMService;
-    private static final Integer PAGE_SIZE = 12;
 
     @Transactional
     public GetRequiredInfoResponse getRequiredInfo(Long userId) {
@@ -93,16 +90,6 @@ public class ProductService {
     }
 
     @Transactional
-    public List<SearchProductResponse> searchProduct
-            (Long userId, Long storeId, String storeConditionValue, String productName) throws IOException {
-        Store store = storeService.findById(storeId);
-        checkRequestIdHasRequestStore(userId, store);
-        StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
-        List<FindProductPage> searchProductUrlList = findProductAllByName(store, storeCondition, productName);
-        return convertUrlToResponse(searchProductUrlList);
-    }
-
-    @Transactional
     public void editProduct(Long userId, Long productId, Product newProduct, MultipartFile image) {
         Product product = findProductById(productId);
         checkRequestIdHasRequestProduct(userId, product);
@@ -117,28 +104,6 @@ public class ProductService {
         Product product = findProductById(productId);
         checkRequestIdHasRequestProduct(userId, product);
         product.delete();
-    }
-
-    @Transactional
-    public List<GetTotalProductResponse> getTotalProduct(Long userId, Long storeId, String storeConditionValue) {
-        Store store = storeService.findById(storeId);
-        checkRequestIdHasRequestStore(userId, store);
-        StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
-        return countProduct(store, storeCondition);
-    }
-
-    @Transactional
-    public List<SearchProductResponse> getListOfSearchProduct
-            (Long userId, Long storeId, String storeConditionValue, String searchConditionValue, Long productId, String sortConditionValue) throws IOException {
-        Product product = configPaging(productId);
-        SortCondition sortCondition = SortCondition.findSortConditionByValue(sortConditionValue);
-        Store store = storeService.findById(storeId);
-        checkRequestIdHasRequestStore(userId, store);
-        StoreCondition storeCondition = StoreCondition.findStoreConditionByValue(storeConditionValue);
-        SearchCondition searchCondition = SearchCondition.findSearchConditionByValue(searchConditionValue);
-        List<FindProductPage> searchProductUrlList = productRepository.findPageOfSearchConditionOrderBySortCondition
-                (store, storeCondition, searchCondition, sortCondition, product.getName(), product.getOrderFreq(), PAGE_SIZE);
-        return convertUrlToResponse(searchProductUrlList);
     }
 
     @Transactional
@@ -171,51 +136,9 @@ public class ProductService {
             throw BaseException.type(ProductErrorCode.DUPLICATE_PRODUCT);
     }
 
-    List<FindProductPage> findProductAllByName(Store store, StoreCondition storeCondition, String productName) {
-        List<FindProductPage> searchProductUrlList = productRepository.findProductByName(store, storeCondition, productName);
-        if (searchProductUrlList.isEmpty())
-            throw BaseException.type(ProductErrorCode.NOT_FOUND_PRODUCT);
-        return searchProductUrlList;
-    }
-
-    private Product findProductById(Long productId) {
+    Product findProductById(Long productId) {
         return productRepository.findProductById(productId)
                 .orElseThrow(() -> BaseException.type(ProductErrorCode.NOT_FOUND_PRODUCT));
-    }
-
-    List<GetTotalProductResponse> countProduct(Store store, StoreCondition storeCondition) {
-        List<GetTotalProductResponse> countList = new ArrayList<>(4);
-        LocalDate currentDate = LocalDate.now();
-        LocalDate standardDate = currentDate.plusDays(3);
-        countList.add(new GetTotalProductResponse
-                ("Total", productRepository.countProductAll(store, storeCondition.getValue())));
-        countList.add(new GetTotalProductResponse
-                ("Pass", productRepository.countProductPass(store, storeCondition.getValue(), currentDate)));
-        countList.add(new GetTotalProductResponse
-                ("Close", productRepository.countProductClose(store, storeCondition.getValue(), currentDate, standardDate)));
-        countList.add(new GetTotalProductResponse
-                ("Lack", productRepository.countProductLack(store, storeCondition.getValue())));
-        return countList;
-    }
-
-    Product configPaging(Long productId) {
-        if (productId == -1)
-            return new Product();
-        return findProductById(productId);
-    }
-
-    private List<SearchProductResponse> convertUrlToResponse(List<FindProductPage> searchProductUrlList) throws IOException {
-        List<SearchProductResponse> searchProductResponseList = new ArrayList<>();
-        for (FindProductPage searchProductUrl : searchProductUrlList) {
-            byte[] image = getImageOrElseNull(searchProductUrl.getImageUrl());
-            SearchProductResponse searchProductResponse = SearchProductResponse.builder()
-                    .id(searchProductUrl.getId())
-                    .name(searchProductUrl.getName())
-                    .image(image)
-                    .build();
-            searchProductResponseList.add(searchProductResponse);
-        }
-        return searchProductResponseList;
     }
 
     byte[] getImageOrElseNull(String imageUrl) throws IOException {
@@ -224,7 +147,7 @@ public class ProductService {
         return fileService.downloadToResponseDto(imageUrl);
     }
 
-    private void checkRequestIdHasRequestStore(Long userId, Store store) {
+    void checkRequestIdHasRequestStore(Long userId, Store store) {
         User user = userFindService.findById(userId);
         if (user.getRole() == Role.SUPERVISOR)
             throw BaseException.type(GlobalErrorCode.INVALID_USER_JWT);
